renamed and moved root macros, unified aligner file names
